requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0

alembic
python-dotenv
bcrypt
PyJWT
psycopg2-binary
email-validator
asyncpg
jose